refactor: define Clock type in Plugin/types
