+++ src/components/molecules/Visualizer/Engine/ref.ts
@@ -2,6 +2,7 @@ import { ComponentType, ReactNode } from "react";
 
 import { LatLngHeight, Camera, Typography, Rect } from "@reearth/util/value";
 
+import { Clock } from "../Plugin/types";
 import type { Component } from "../Primitive";
 
 export type MouseEvent = {
@@ -46,19 +47,6 @@ export type MouseEventHandles = {
   onMouseEnter: (fn: MouseEvents["mouseenter"]) => void;
   onMouseLeave: (fn: MouseEvents["mouseleave"]) => void;
   onWheel: (fn: MouseEvents["wheel"]) => void;
-};
-
-type ClockEventHandler<R = void> = (cb: (clock: Clock) => void) => R;
-export type Clock = {
-  startTime: Date;
-  stopTime: Date;
-  currentTime: Date;
-  tick: () => Date;
-  shouldAnimate: boolean;
-  onTick: {
-    addEventListener: ClockEventHandler<() => void>;
-    removeEventListener: ClockEventHandler;
-  };
 };
 
 export type EngineRef = {
