+++ src/components/molecules/Visualizer/Plugin/context.tsx
@@ -10,7 +10,7 @@ import {
 import events from "@reearth/util/event";
 import { Rect } from "@reearth/util/value";
 
-import { MouseEvents, MouseEventHandles, Clock } from "../Engine/ref";
+import { MouseEvents, MouseEventHandles } from "../Engine/ref";
 import type { LayerStore } from "../Layers";
 import type { Component as PrimitiveComponent } from "../Primitive";
 import { useGet } from "../utils";
@@ -26,6 +26,7 @@ import type {
   LookAtDestination,
   Tag,
   MouseEvent,
+  Clock,
 } from "./types";
 
 export type EngineContext = {
